[minor] Merge lolcation into main module
